+++ src/api/endpoints/images/__tests__/images-api.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect, vi, beforeEach } from 'vitest';
 import { imagesApi, type ImageApiResponse, type ImageApiOptions } from '../index';
 import type { HttpClient } from '../../../client';
-import type { DeepSnakeCase } from '../../../../utils/type-transformers';
+import type { DeepSnakeCase } from '../../../../utils/case-converter/index.js';
 import { TestData } from '../../../../constants';
 
 describe('imagesApi', () => {
