+++ src/api/endpoints/images/index.ts
@@ -1,7 +1,7 @@
 // 画像エクスポート関連のAPI呼び出し関数
 
 import type { HttpClient } from '../../client.js';
-import type { DeepSnakeCase } from '../../../utils/type-transformers.js';
+import type { DeepSnakeCase } from '../../../utils/case-converter/index.js';
 import { buildUrlParams } from '../utils/params-builder.js';
 
 // API Options
